chunk generation based in world size

+++ src/render/model.rs
@@ -1,11 +1,10 @@
-
 use crate::render::{buffer::Buffer, mesh::Mesh};
 
 use super::Vertex;
 /// Represents a mesh that has been sent to the GPU.
 pub struct Model<V: Vertex>{
     vbuf: Buffer<V>,
-    ibuf: Buffer<u16>,
+    ibuf: Buffer<u32>,
     pub num_indices: u32,
 }
